+++ src/Sdcb.PaddleDetection/TensorOutput.cs
@@ -6,7 +6,7 @@ namespace Sdcb.PaddleDetection;
 internal struct TensorOutput
 	{
 		public int[] Shape;
-		public DataTypes DataType;
+		public PaddleDataType DataType;
 		public string Name;
 		public Array Data;
 	}
