+++ src/Sdcb.PaddleInference/PaddleConfig.cs
@@ -2,7 +2,6 @@
 using Sdcb.PaddleInference.TensorRt;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -243,7 +242,7 @@ public sealed class PaddleConfig : IDisposable
         PaddlePrecision precision = PaddlePrecision.Float32,
         bool useStatic = true,
         bool useCalibMode = false)
-        => PaddleNative.PD_ConfigEnableTensorRtEngine(_ptr, workspaceSize, maxBatchSize, minSubgraphSize, (int)precision, (sbyte)(useStatic ? 1 : 0), (sbyte)(useCalibMode ? 1 : 0));
+        => PaddleNative.PD_ConfigEnableTensorRtEngine(_ptr, workspaceSize, maxBatchSize, minSubgraphSize, precision, (sbyte)(useStatic ? 1 : 0), (sbyte)(useCalibMode ? 1 : 0));
 
     /// <summary>A boolean state telling whether the TensorRT engine is used.</summary>
     public bool TensorRtEngineEnabled => PaddleNative.PD_ConfigTensorRtEngineEnabled(_ptr) != 0;
@@ -352,9 +351,9 @@ public sealed class PaddleConfig : IDisposable
     public bool Valid => PaddleNative.PD_ConfigIsValid(_ptr) != 0;
 
     /// <summary>Turn on GPU.</summary>
-    public void EnableUseGpu(int initialMemoryMB, int deviceId)
+    public void EnableUseGpu(int initialMemoryMB, int deviceId, PaddlePrecision precision = PaddlePrecision.Float32)
     {
-        PaddleNative.PD_ConfigEnableUseGpu(_ptr, (ulong)initialMemoryMB, deviceId);
+        PaddleNative.PD_ConfigEnableUseGpu(_ptr, (ulong)initialMemoryMB, deviceId, precision);
     }
 
     /// <summary>Set the combined model with two specific pathes for program and parameters.</summary>
@@ -411,7 +410,15 @@ public sealed class PaddleConfig : IDisposable
         }
     }
 
-    /// <summary>Delete all passes that has a certain type 'pass'.</summary>
+    /// <summary>
+    /// Deletes a pass with the specified name.
+    /// </summary>
+    /// <param name="passName">The name of the pass to be deleted.</param>
+    /// <remarks>
+    /// This method deletes a pass from the configuration using the provided name.
+    /// </remarks>
+    /// <exception cref="ArgumentNullException">Thrown if the <paramref name="passName"/> is null.</exception>
+    /// <seealso cref="PaddleNative.PD_ConfigDeletePass(IntPtr, IntPtr)"/>
     public unsafe void DeletePass(string passName)
     {
         byte[] passNameBytes = PaddleEncoding.GetBytes(passName);
@@ -440,63 +447,5 @@ public sealed class PaddleConfig : IDisposable
     {
         configure(this);
         return this;
-    }
-}
-
-class PtrFromStringArray : IDisposable
-{
-    readonly IntPtr[] internalArray;
-    readonly GCHandle[] handles;
-    readonly GCHandle mainHandle;
-
-    public PtrFromStringArray(string[] data)
-    {
-        handles = new GCHandle[data.Length];
-        internalArray = new IntPtr[data.Length];
-
-        for (int i = 0; i < data.Length; ++i)
-        {
-            byte[] byteArray = Encoding.UTF8.GetBytes(data[i] + '\0');
-            handles[i] = GCHandle.Alloc(byteArray, GCHandleType.Pinned);
-            internalArray[i] = handles[i].AddrOfPinnedObject();
-        }
-
-        mainHandle = GCHandle.Alloc(internalArray, GCHandleType.Pinned);
-    }
-
-    public IntPtr Ptr => mainHandle.AddrOfPinnedObject();
-
-    public void Dispose()
-    {
-        foreach (GCHandle handle in handles) handle.Free();
-        mainHandle.Free();
-    }
-}
-
-class PtrFromIntArray : IDisposable
-{
-    readonly IntPtr[] internalArray;
-    readonly GCHandle[] handles;
-    readonly GCHandle mainHandle;
-
-    public PtrFromIntArray(int[][] data)
-    {
-        handles = new GCHandle[data.Length];
-        internalArray = new IntPtr[data.Length];
-        mainHandle = GCHandle.Alloc(internalArray, GCHandleType.Pinned);
-
-        for (int i = 0; i < data.Length; ++i)
-        {
-            handles[i] = GCHandle.Alloc(data[i], GCHandleType.Pinned);
-            internalArray[i] = handles[i].AddrOfPinnedObject();
-        }
-    }
-
-    public IntPtr Ptr => mainHandle.AddrOfPinnedObject();
-
-    public void Dispose()
-    {
-        foreach (GCHandle handle in handles) handle.Free();
-        mainHandle.Free();
     }
 }
