+++ src/Sdcb.PaddleInference/Sdcb.PaddleInference.csproj
@@ -3,7 +3,7 @@
     <TargetFrameworks>net6;netstandard2.0;net45</TargetFrameworks>
     <Nullable>enable</Nullable>
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <LangVersion>10</LangVersion>
+    <LangVersion>preview</LangVersion>
     <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
     <PackageProjectUrl>https://github.com/sdcb/PaddleSharp</PackageProjectUrl>
     <RepositoryUrl>https://github.com/sdcb/PaddleSharp.git</RepositoryUrl>
